Moved twig integrations to correct namespace
